requirements.txt
numpy==1.24.3
matplotlib==3.7.1
PyYAML==6.0
tqdm==4.65.0
torch==2.0.1
Pillow==9.5.0
